Parse options in boombox bin as in boombox script
